Fix tooling and relax searchProducts
