+++ packages/api/src/mcp/tools/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Re-export all MCP tools
+ */
+
+export * from './listProducts.js';
